requirements.txt
pandas
streamlit

folium
plotly-express
streamlit-folium
